backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
supabase>=2.0.0
python-multipart>=0.0.9
